backend/requirements.txt
aiosqlite
asyncpg
fastapi
uvicorn
sqlalchemy
pydantic
langgraph
llama-index
redis
python-dotenv
websockets
psycopg2-binary
alembic
